First time data grab

+++ server/schema.ts
@@ -4,7 +4,7 @@ const schema = buildSchema(`
    type Item {
     id: ID!
     title: String!
-    price: Int!
+    price: Float!
     pictureUri: String
     stockAmount: Int!
   }
@@ -15,8 +15,8 @@ const schema = buildSchema(`
     }
 
   type Mutation {
-        createItem(title: String!, price: Int!, pictureUri: String, stockAmount: Int!): Item!
-        updateItem(id: ID!, title: String, price: Int, pictureUri: String, stockAmount: Int): Item!
+        createItem(title: String!, price: Float!, pictureUri: String, stockAmount: Int!): Item!
+        updateItem(id: ID!, title: String, price: Float, pictureUri: String, stockAmount: Int): Item!
         deleteItem(id: ID!): ID!
     }
 `);
